Refactor application form components for improved modularity and functionality
- Reorganized the application form into multiple dedicated components: `BasicInfoSection`, `TextAreaSection`, `ContributionAreasSection`, `HowKnowUsSection`, and `OptionalFieldsSection` to enhance maintainability and readability.
- Introduced custom hooks `useApplicationForm` and `useFormSelections` for better state management and form handling.
- Updated validation schema to ensure proper handling of form submissions and user inputs.
- Enhanced UI with improved styling and user feedback mechanisms, including error messages and loading states.
- Removed the legacy `ApplicationForm` component to streamline the codebase.
